// frontend/src/app/api/notes/model/noteDto.ts
/**
 * PaperNote Notes API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Complete note information including content, metadata, tags and sharing details
 */
export interface NoteDto { 
    /**
     * Unique identifier of the note
     */
    id?: string;
    /**
     * Title of the note
     */
    title?: string | null;
    /**
     * Full content of the note
     */
    content?: string | null;
    /**
     * Date and time when the note was created
     */
    createdAt?: string;
    /**
     * Date and time when the note was last updated
     */
    updatedAt?: string;
    /**
     * List of tags associated with the note
     */
    tags?: Array<string> | null;
    /**
     * Username of the note owner
     */
    ownerUsername?: string | null;
    /**
     * List of usernames the note is shared with
     */
    sharedWithUsernames?: Array<string> | null;
}