// frontend/src/app/api/notes/model/noteSummaryDto.ts
/**
 * PaperNote Notes API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Condensed note information for list views with preview content
 */
export interface NoteSummaryDto { 
    /**
     * Unique identifier of the note
     */
    id?: string;
    /**
     * Title of the note
     */
    title?: string | null;
    /**
     * Preview of the note content (truncated)
     */
    contentPreview?: string | null;
    /**
     * Date and time when the note was created
     */
    createdAt?: string;
    /**
     * Date and time when the note was last updated
     */
    updatedAt?: string;
    /**
     * List of tags associated with the note
     */
    tags?: Array<string> | null;
    /**
     * Username of the note owner
     */
    ownerUsername?: string | null;
}